```
hyperfine --warmup 3 "/usr/bin/fold -w 40 moby64.txt"                     "./target/release/coreutils.prev fold -w 40 moby64.txt" "./target/release/coreutils fold -w 40 moby64.txt"
Benchmark 1: /usr/bin/fold -w 40 moby64.txt
  Time (mean ± σ):     167.7 ms ±   3.4 ms    [User: 154.7 ms, System: 12.8 ms]
  Range (min … max):   162.7 ms … 174.8 ms    17 runs

Benchmark 2: ./target/release/coreutils.prev fold -w 40 moby64.txt
  Time (mean ± σ):     521.3 ms ±   8.3 ms    [User: 347.4 ms, System: 173.8 ms]
  Range (min … max):   516.1 ms … 543.2 ms    10 runs

Benchmark 3: ./target/release/coreutils fold -w 40 moby64.txt
  Time (mean ± σ):     167.0 ms ±   2.6 ms    [User: 156.9 ms, System: 9.9 ms]
  Range (min … max):   162.8 ms … 170.7 ms    17 runs

Summary
  ./target/release/coreutils fold -w 40 moby64.txt ran
    1.00 ± 0.03 times faster than /usr/bin/fold -w 40 moby64.txt
    3.12 ± 0.07 times faster than ./target/release/coreutils.prev fold -w 40 moby64.txt
```